game_save.cpp: More warning fixes

+++ src/main/game_save.cpp
@@ -131,7 +131,7 @@ void FindSaves()
                 curActive++;
 
             // How much paths open
-            maxActive += f.visiblePaths.size();
+            maxActive += (int)f.visiblePaths.size();
             for(auto &p : f.visiblePaths)
             {
                 if(p.second)
@@ -139,7 +139,7 @@ void FindSaves()
             }
 
             // How much levels opened
-            maxActive += f.visibleLevels.size();
+            maxActive += (int)f.visibleLevels.size();
             for(auto &p : f.visibleLevels)
             {
                 if(p.second)
